chore: load the deploed model
